stick-controlled gain uses the VR direction to set polarity of setpoint.
the stick position (direction and magnitude) is used to set the pid setpoint. need to set the polarity to be canceled out by the VR polarity later. note that there might be an issue with vr magnitude is close to zero, so the sign can switch between +1 and -1, but since the VR magnitude is also close to zero, the correction will be scaled down to zero anyways.