Improve one-wire protocol by ensuring the response type matches request type.
on erase-stats (OW_SET_STATS) or erase-cfg (OW_SET_CFG), the rx3s may take a long time to respond (writing to eeprom), so the program box times out waiting for the response and restarts the state machine. when the OW_SET_STATS/OW_SET_CFG response finally arrives, the program box thinks that it is for the OW_GET_CFG in the WAIT_CONFIG state, and mis-parses the bad response.

the change is to enaure that the response type must match the request time (in this case OW_GET_CFG), particular on session/state machine restarts.