bug fix for rx frame sync and servo frame update
rx frames are updated
1. high rate with serial bus rx (spektrum/sbus)
2. normal rate with regular or cppm rx
3. on timeout (30ms)

ensure that servo frames are not updated at the same rate as the rx on (1a), but instead rate limited to minimum_servo_frame_time.

this minimum_servo_frame_time parameter is set to
- 18ms when serialrx_spektrum/sbus is defined
- 0ms otherwise.

this will allow (1b) to update the servos as fast as the rx will generate them. assumption is that you can set the rx rate to LO/HI (11ms/22ms for example) to support either analog or digital servos.

we also assume that the timeout in (1c) is high, ie: maximum_rx_frame_time >> minimum_servo_frame_time
(in this future, we may use this for failsafe)